refactor/examples: move examples into independent crates, move generic code to juice-utils
