+++ juice-examples/mackey-glass-rnn-regression/assets/.gitignore
@@ -1,0 +1,2 @@
+*
+!.gitignore
