+++ juice-examples/mnist-image-multiclass-classification/.gitignore
@@ -1,0 +1,1 @@
+target
