+++ juice-examples/mnist-image-multiclass-classification/Cargo.toml
@@ -1,0 +1,40 @@
+[package]
+name = "example-mnist-classification"
+description = "MNIST Image Multiclass Classification example of Juice"
+version = "0.0.1"
+authors = ["Bernhard Schuster <bernhard@ahoi.io>",
+           "Michael Hirn <mj@autumnai.com>",
+           "Maximilian Goisser <max@autumnai.com>"]
+edition = "2018"
+
+[dependencies]
+greenglas = { version = "0.2" }
+juice = { version = "0.2.4", default-features = false }
+coaster = { version = "0.1.1", default-features = false }
+coaster-nn = { version = "0.4.2", default-features = false }
+
+juice-utils = {path = "../juice-utils"}
+
+csv = "1"
+hyper = "0.13"
+hyper-rustls = "0.20"
+futures = "0.3"
+futures-util = "0.3"
+
+tokio = "0.2"
+flate2 = "1"
+mnist = "0.4.0"
+
+log = "0.4"
+env_logger = "0.7"
+timeit = "0.1.2"
+
+docopt = "1"
+serde = { version = "1", features = ["derive"] }
+
+
+[features]
+default = ["native", "cuda"]
+native = ["juice/native", "coaster/native"]
+opencl = ["juice/opencl", "coaster/opencl"]
+cuda = ["juice/cuda", "coaster/cuda"]
