+++ juice-examples/mnist-image-multiclass-classification/assets/.gitignore
@@ -1,0 +1,2 @@
+*
+!.gitignore
